DRTII-500 Fix expected recovery message type

+++ server/src/main/scala/actors/migration/FlightsMigrationActor.scala
@@ -75,7 +75,7 @@ class FlightsMigrationActor(journalType: StreamingJournalLike, flightMigrationRo
     case RecoveryCompleted =>
       log.info(s"Recovered migration at ${state.seqNr} sequence number")
 
-    case recoveredLastProcessed: Int =>
+    case recoveredLastProcessed: Long =>
       state = state.copy(seqNr = recoveredLastProcessed)
 
     case unexpected => log.error(s"Received unexpected recovery message: $unexpected")
